Move the toeplitz solve tests to a new file.
This may help evaluate how much time a given test will take.
